+++ Apps/macOS/PHTV/Utilities/PHTVObjectiveCBridge.h
@@ -1,0 +1,6 @@
+//
+//  PHTVObjectiveCBridge.h
+//  PHTV
+//
+
+#import "../System/PHTVRemoteViewCrashGuard.h"